feat: automate sitemap generation
