+++ src/components/data/pageSeo.js
@@ -1,0 +1,377 @@
+export const pageSeo = {
+  Home: {
+    title: 'UK Salary, Tax & Mortgage Calculators | Calculate My Money',
+    description:
+      'Free UK calculators for take-home pay, tax, NI, mortgages, savings and ROI. Fast, accurate tools for 2025/26.',
+  },
+  Resources: {
+    title: 'Financial Resources & Guides | Calculate My Money',
+    description:
+      'Explore our financial resources and guides on Calculate My Money. Learn about tax, savings, investments, and personal finance in the UK.',
+  },
+  Blog: {
+    title: 'Financial Blog | Insights & Articles - Calculate My Money',
+    description:
+      'Read the latest articles and insights on UK personal finance, tax, budgeting, and investment strategies from the Calculate My Money blog.',
+  },
+  BlogSmartMoneySavingTips: {
+    title: 'Smart Money-Saving Tips for UK Families | Calculate My Money',
+    description:
+      'Discover practical tips for UK families to save money on groceries, energy bills, and everyday expenses. A complete guide to cut costs and budget effectively.',
+  },
+  BlogDebtRepaymentStrategies: {
+    title: 'Debt Snowball vs. Avalanche in the UK | Calculate My Money',
+    description:
+      'Compare the Debt Snowball and Debt Avalanche methods. Find the best strategy to pay off your debts in the UK, from credit cards to loans.',
+  },
+  BlogFinancialPsychology: {
+    title: 'Your Relationship with Money: A Guide to Financial Psychology',
+    description:
+      'Understand the psychology behind your spending and saving habits. Learn how your money mindset impacts your financial health and future prosperity in the UK.',
+  },
+  Contact: {
+    title: 'Contact Us | Calculate My Money',
+    description:
+      "Get in touch with Calculate My Money for support, feedback, or business inquiries. We're here to help with your financial calculator needs.",
+  },
+  PrivacyPolicy: {
+    title: 'Privacy Policy | Calculate My Money',
+    description:
+      'Understand our Privacy Policy at Calculate My Money. Learn how we collect, use, and protect your personal data.',
+  },
+  CookiePolicy: {
+    title: 'Cookie Policy | Calculate My Money',
+    description:
+      'Review our Cookie Policy at Calculate My Money to understand how we use cookies on our website.',
+  },
+  TermsOfService: {
+    title: 'Terms of Service | Calculate My Money',
+    description:
+      'Read the Terms of Service for Calculate My Money. By using our website, you agree to these terms.',
+  },
+  Disclaimer: {
+    title: 'Disclaimer | Calculate My Money',
+    description:
+      'Important disclaimer regarding the use of financial calculators and information provided on Calculate My Money.',
+  },
+  Sitemap: {
+    title: 'Sitemap | Calculate My Money',
+    description:
+      'Browse the sitemap for Calculate My Money to find all available financial calculators and resources.',
+  },
+  SalaryCalculatorUK: {
+    title: 'UK Salary Calculator (Take-Home Pay 2025/26)',
+    description:
+      'Calculate your UK take-home pay after tax and NI for 2025/26. Includes paycheck, pro-rata and gross-to-net tools.',
+  },
+  SalaryCalculatorTakeHomePay: {
+    title: 'Take-Home Pay Calculator UK (2025/26)',
+    description: 'Estimate UK take-home pay after tax, NI, pension & student loans for 2025/26.',
+  },
+  SalaryCalculatorPaycheck: {
+    title: 'Paycheck Calculator UK (After Tax & NI)',
+    description:
+      'Work out your UK paycheck after tax and NI. Supports weekly, fortnightly and monthly pay.',
+  },
+  GrossToNetCalculator: {
+    title: 'Gross to Net Income Calculator UK',
+    description: 'Convert gross salary to net take-home pay with UK tax and NI for 2025/26.',
+  },
+  ProRataSalaryCalculator: {
+    title: 'Pro-Rata Salary Calculator UK',
+    description: 'Convert annual salary to pro-rata pay by hours or days worked in the UK.',
+  },
+  MortgageCalculator: {
+    title: 'UK Mortgage Calculator | Repayments & Affordability - Calculate My Money',
+    description:
+      'Estimate your UK mortgage repayments and check affordability with our free mortgage calculator. See monthly costs, interest, and the full amortization schedule.',
+  },
+  BudgetCalculator: {
+    title: 'Budget Planner UK | Free Online Budgeting Tool - Calculate My Money',
+    description:
+      'Create a personal budget with our free UK budget planner. Track income and expenses to manage your finances effectively.',
+  },
+  CompoundInterestCalculator: {
+    title: 'Compound Interest Calculator UK | Grow Your Savings - Calculate My Money',
+    description:
+      'See how your savings can grow with our UK compound interest calculator. Plan for future investments and financial goals.',
+  },
+  PensionCalculator: {
+    title: 'UK Pension Calculator | Plan Your Retirement - Calculate My Money',
+    description:
+      'Estimate your future pension income with our UK pension calculator. Plan for a secure retirement.',
+  },
+  NationalInsuranceCalculator: {
+    title: 'UK National Insurance Calculator 2025/26 | Calculate Your NI',
+    description:
+      "Calculate your UK National Insurance contributions for the 2025/26 tax year. Our free NI calculator shows you exactly what you'll pay based on your salary.",
+  },
+  MaternityPayCalculator: {
+    title: 'Statutory Maternity Pay Calculator UK 2025 | SMP Estimator',
+    description:
+      "Estimate your Statutory Maternity Pay (SMP) with our free UK calculator. See how much you'll receive for the 39 weeks of your maternity leave.",
+  },
+  SeverancePayCalculator: {
+    title: 'UK Severance Pay Calculator | Redundancy & Exit Package Estimate',
+    description:
+      'Estimate UK severance pay including statutory redundancy, notice pay, unused holiday and contractual enhancements.',
+  },
+  InflationCalculator: {
+    title: 'UK Inflation Calculator | Calculate the Changing Value of Money',
+    description:
+      'See how the value of the pound has changed over time with our free UK inflation calculator. Compare the purchasing power of money between any two years.',
+  },
+  IncomeTaxCalculator: {
+    title: 'UK Income Tax Calculator 2025/26 | Calculate Your Tax Bill - Calculate My Money',
+    description:
+      'Calculate your income tax for 2025/26 with our free UK tax calculator. Covers tax bands for England, Scotland, Wales & NI. Find out your true tax liability.',
+  },
+  SalaryIncreaseCalculator: {
+    title: 'Salary Increase Calculator UK | See Your New Pay - Calculate My Money',
+    description:
+      'Calculate how a percentage-based pay rise will affect your annual and monthly gross income with our simple salary increase calculator.',
+  },
+  SimpleInterestCalculator: {
+    title: 'Simple Interest Calculator UK | Calculate My Money',
+    description:
+      'Use our simple interest calculator to quickly determine the interest earned on a principal sum over a fixed period without compounding.',
+  },
+  EffectiveTaxRateCalculator: {
+    title: 'Effective Tax Rate Calculator UK | Your True Tax Rate - Calculate My Money',
+    description:
+      'Find your true, overall tax burden. Our calculator shows your effective tax rate after factoring in your personal allowance, income tax, and NI.',
+  },
+  HomeEquityLoanCalculator: {
+    title: 'Home Equity Loan Calculator UK | How Much Can You Borrow? - Calculate My Money',
+    description:
+      'Estimate how much equity you can borrow from your home with our free UK home equity loan calculator. Understand your LTV and available funds.',
+  },
+  AnnuityCalculator: {
+    title: 'Annuity Calculator UK | Estimate Your Retirement Income - Calculate My Money',
+    description:
+      'Use our annuity calculator to estimate the guaranteed income you could receive from your pension pot for a fixed term or for life.',
+  },
+  CommissionCalculator: {
+    title: 'Commission Calculator UK | Calculate Sales Earnings - Calculate My Money',
+    description:
+      'Easily calculate your gross commission earnings based on sales revenue and commission percentage with our free online tool.',
+  },
+  DividendTaxCalculator: {
+    title: 'UK Dividend Tax Calculator 2024/25 | Calculate My Money',
+    description:
+      'Estimate your tax liability on dividend income for the 2024/25 tax year. Accounts for allowances and income tax bands.',
+  },
+  FutureValueCalculator: {
+    title: "Future Value Calculator | Project Your Investment's Worth - Calculate My Money",
+    description:
+      'Calculate the future value of a single lump-sum investment based on a constant interest rate over a specific period.',
+  },
+  OvertimePayCalculator: {
+    title: 'Overtime Pay Calculator UK | Calculate Your Extra Earnings - Calculate My Money',
+    description:
+      'Work out your gross pay including overtime hours. Set your standard rate, overtime hours, and pay multiplier to see your total earnings.',
+  },
+  LoanComparisonCalculator: {
+    title: 'Loan Comparison Calculator UK | Find The Better Deal - Calculate My Money',
+    description:
+      'Compare two loan offers side-by-side to determine the true cost. Analyze monthly payments, total interest, and the total amount repaid.',
+  },
+  InheritanceTaxCalculator: {
+    title: 'Inheritance Tax Calculator UK | IHT Estimator - Calculate My Money',
+    description:
+      'Estimate your potential Inheritance Tax (IHT) liability with our easy-to-use UK calculator. Understand the nil-rate bands.',
+  },
+  CouncilTaxCalculator: {
+    title: 'Council Tax Calculator UK | Estimate Your Bill - Calculate My Money',
+    description:
+      "Get an estimate of your annual Council Tax bill based on your property's band. Covers average rates for England.",
+  },
+  MortgageRepaymentCalculator: {
+    title: 'Mortgage Repayment Schedule Calculator | Amortization - Calculate My Money',
+    description:
+      'Generate a full mortgage repayment (amortization) schedule. See the breakdown of principal and interest and interest for each payment.',
+  },
+  FirstTimeBuyerCalculator: {
+    title: 'First-Time Buyer Calculator UK | Affordability & Costs - Calculate My Money',
+    description:
+      'Assess your affordability as a first-time buyer. Calculate your loan-to-income ratio, deposit percentage, and estimated stamp duty.',
+  },
+  RentVsBuyCalculator: {
+    title: 'Rent vs Buy Calculator UK | Financial Comparison - Calculate My Money',
+    description:
+      'Compare the estimated monthly costs of renting a property versus buying a home to help you make an informed financial decision.',
+  },
+  HouseholdBillsSplitter: {
+    title: 'Household Bill Splitter Calculator | Calculate My Money',
+    description:
+      'Easily split rent, utilities, and other shared household expenses between multiple people with our simple bill splitting tool.',
+  },
+  CommuteCostCalculator: {
+    title: 'Commute Cost Calculator UK | Fuel & Transport Costs - Calculate My Money',
+    description:
+      'Calculate the weekly and monthly cost of your commute, factoring in fuel consumption, prices, and public transport expenses.',
+  },
+  CarCostCalculator: {
+    title: 'True Cost of Car Ownership Calculator UK | Calculate My Money',
+    description:
+      'Calculate the total cost of owning a car, including depreciation, fuel, insurance, tax, and maintenance over your ownership term.',
+  },
+  SubscriptionCostCalculator: {
+    title: 'Subscription Cost Calculator | Track Your Spending - Calculate My Money',
+    description:
+      'Track all your monthly and annual subscriptions in one place to see the total cost and find areas to save money.',
+  },
+  RuleOf72Calculator: {
+    title: 'Rule of 72 Calculator | Doubling Time for Investments - Calculate My Money',
+    description:
+      'Use the Rule of 72 to quickly estimate how many years it will take for your investment to double at a given rate of return.',
+  },
+  StudentLoanRepaymentCalculator: {
+    title: 'Student Loan Repayment Calculator UK (All Plans) | Calculate My Money',
+    description:
+      'Calculate your estimated monthly and annual student loan repayments based on your salary and specific UK loan plan (Plan 1, 2, 4, 5, or Postgraduate).',
+  },
+  WeddingBudgetCalculator: {
+    title: 'Wedding Budget Calculator & Tracker | Calculate My Money',
+    description:
+      'Plan and track your wedding expenses with our comprehensive budget calculator. Manage your spending across all categories.',
+  },
+  TravelBudgetCalculator: {
+    title: 'Travel & Holiday Budget Calculator | Calculate My Money',
+    description:
+      'Plan your next holiday with our easy-to-use travel budget calculator. Estimate costs for flights, accommodation, food, and activities.',
+  },
+  ChildcareCostCalculator: {
+    title: 'Childcare Cost Calculator UK | Nursery Fees Estimator - Calculate My Money',
+    description:
+      'Estimate the weekly, monthly, and annual cost of childcare based on daily rates and the number of days required.',
+  },
+  ChildBenefitCalculator: {
+    title: 'UK Child Benefit Calculator | High Income Charge & Payments',
+    description:
+      'Check your weekly Child Benefit payments, see any High Income Child Benefit Charge and estimate how much you keep.',
+  },
+  TipCalculator: {
+    title: 'UK Tip & Bill Splitting Calculator | Calculate My Money',
+    description:
+      'Easily calculate a tip and split the total bill between any number of people. Perfect for dining out in the UK.',
+  },
+  OvertimeRateCalculator: {
+    title: 'Overtime Rate Calculator UK | Calculate My Money',
+    description:
+      'Calculate your hourly pay rate for overtime work based on your standard rate and contract multiplier (e.g., time-and-a-half).',
+  },
+  CurrencyConverter: {
+    title: 'Currency Converter & Exchange Rate Hub | Calculate My Money',
+    description:
+      'Understand currency exchange and find links to reliable, live exchange rates for GBP, USD, EUR, and more. An educational tool for travellers and investors.',
+  },
+  UKGovernmentBudget: {
+    title: 'UK Government Budget Analysis | Calculate My Money',
+    description:
+      'Detailed analysis of the latest UK Government Budget announcements and their impact on your finances and calculators.',
+  },
+  UKFinancialStats: {
+    title: 'Live UK Financial Statistics Dashboard | BoE Rate, Inflation, House Prices',
+    description:
+      'Track key UK economic indicators in real-time. Our dashboard provides the latest Bank of England interest rate, inflation (CPI), house prices, and more.',
+  },
+  JobSalaries: {
+    title: 'Average Job Salaries UK | By Industry & Region - Calculate My Money',
+    description:
+      'Discover average job salaries across various industries and regions in the UK. Compare earnings and career prospects with our tools.',
+  },
+  CostOfLiving: {
+    title: 'Cost of Living UK | City & Regional Breakdown - Calculate My Money',
+    description:
+      'Explore the cost of living in various UK cities and regions. Understand typical expenses for housing, food, and transport with our data.',
+  },
+  Methodology: {
+    title: 'Methodology & Data Sources (HMRC, BoE) | Calculate My Money',
+    description:
+      'How we calculate UK salary, tax and mortgage results. Data sources: HMRC 2025/26, Bank of England.',
+  },
+  About: {
+    title: 'About Calculate My Money – UK Financial Calculators',
+    description: 'Who we are and how we build accurate, transparent UK financial calculators.',
+  },
+  TaxCalculatorsUK: {
+    title: 'UK Tax Calculators Hub | Income Tax, NI, Net Income',
+    description:
+      'Explore UK tax tools for 2025/26: Income Tax after tax, Tax + NI, Net Income and Self Assessment guides.',
+  },
+  TaxAfterTaxCalculator: {
+    title: 'Tax After Tax Calculator UK | 2025/26',
+    description:
+      'Work out your UK tax after tax for the 2025/26 year. Clear band-by-band breakdowns.',
+  },
+  TaxAndNICalculator: {
+    title: 'Tax and NI Calculator UK | Combined Deductions 2025/26',
+    description: 'Calculate combined UK Income Tax and National Insurance for 2025/26.',
+  },
+  NetIncomeUKCalculator: {
+    title: 'Net Income Calculator UK | Take-Home After Tax & NI',
+    description: 'Estimate your UK net income after tax and NI deductions for 2025/26.',
+  },
+  SelfAssessmentGuide: {
+    title: 'UK Self Assessment Guide | Deadlines, Rates, Tips',
+    description:
+      'Understand UK Self Assessment: deadlines, allowances, rates and tips for 2025/26.',
+  },
+  MortgageCalculatorUK: {
+    title: 'UK Mortgage Calculators Hub | Repayments, Comparison',
+    description:
+      'Explore UK mortgage tools: loan repayment, comparisons, and home loan calculators for 2025/26.',
+  },
+  MortgageLoanRepayment: {
+    title: 'Mortgage Loan Repayment Calculator UK',
+    description: 'Estimate monthly mortgage repayments and total interest across the term.',
+  },
+  HomeLoanMortgageCalculator: {
+    title: 'Home Loan Mortgage Calculator UK',
+    description: 'Quick home loan mortgage estimates: payments by rate, term and deposit.',
+  },
+  MortgageComparison: {
+    title: 'Mortgage Comparison Calculator UK',
+    description: 'Compare two mortgage deals side-by-side. See total costs and savings.',
+  },
+  // Add entries for other specific pages/calculators here following the same pattern
+};
+
+// ---- OG/Twitter images for key salary pages (merge without replacing other entries) ----
+const CDN = 'https://xifmvsuddgebmlleggqz.supabase.co/storage/v1/object/public/CalcMyMoney.co.uk';
+
+pageSeo.SalaryCalculatorUK = {
+  ...pageSeo.SalaryCalculatorUK,
+  ogImage: `${CDN}/og-final-salary-hub.png`,
+  ogImageAlt: 'UK Salary Calculator – Take-Home Pay (2025/26)',
+};
+
+pageSeo.SalaryCalculatorTakeHomePay = {
+  ...pageSeo.SalaryCalculatorTakeHomePay,
+  ogImage: `${CDN}/og-final-take-home.png`,
+  ogImageAlt: 'Take-Home Pay Calculator UK (2025/26)',
+};
+
+pageSeo.SalaryCalculatorPaycheck = {
+  ...pageSeo.SalaryCalculatorPaycheck,
+  ogImage: `${CDN}/og-final-paycheck.png`,
+  ogImageAlt: 'Paycheck Calculator UK – After Tax & NI',
+};
+
+pageSeo.GrossToNetCalculator = {
+  ...pageSeo.GrossToNetCalculator,
+  ogImage: `${CDN}/og-final-gross-to-net.png`,
+  ogImageAlt: 'Gross to Net Income Calculator (UK)',
+};
+
+pageSeo.ProRataSalaryCalculator = {
+  ...pageSeo.ProRataSalaryCalculator,
+  ogImage: `${CDN}/og-final-pro-rata.png`,
+  ogImageAlt: 'Pro-Rata Salary Calculator (UK)',
+};
+
+// Add a default Open Graph image (used if page-specific image isn't set)
+export const defaultOgImage =
+  'https://xifmvsuddgebmlleggqz.supabase.co/storage/v1/object/public/CalcMyMoney.co.uk/og-default.png';
+export const defaultOgAlt = 'Calculate My Money – Free UK Calculators';
